+++ src/test/java/TestWorkspace.java
@@ -95,6 +95,4 @@ public class TestWorkspace {
             throw new Exception("ids.size() != 4");
         }
     }
-
-
 }