+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ### Added
 - Event.workspaceId property to multiple Workspace Webhook identification
 - Workspace.update() to allow parameter updates
+- Transfer.description property to allow control over corresponding Transaction descriptions
 
 ### Fixed
 - special characters in brcodePreview query